add external link attrs to markdown links
