--- cfg/UrlFiltersPlugin.mjs
@@ -1,9 +1,0 @@
-export function UrlFiltersPlugin(config) {
-  config.addFilter("toExternalLinkAttrs", (url) => {
-    if (url.startsWith("http")) {
-      return `target="_blank" rel="noopener noreferrer"`
-    }
-
-    return undefined
-  })
-}
