+++ cfg/index.mjs
@@ -1,8 +1,8 @@
 export { CollectionsPlugin } from "./CollectionsPlugin.mjs"
 export { DateFiltersPlugin } from "./DateFiltersPlugin.mjs"
-export { UrlFiltersPlugin } from "./UrlFiltersPlugin.mjs"
+export { ExternalLinksPlugin } from "./ExternalLinksPlugin.mjs"
 export { ImagePlugin } from "./ImagePlugin.mjs"
 export { RandPlugin } from "./RandPlugin.mjs"
+export { RemoveTrailingSlashPlugin } from "./RemoveTrailingSlashPlugin.mjs"
 export { SassPlugin } from "./SassPlugin.mjs"
 export { StringFiltersPlugin } from "./StringFiltersPlugin.mjs"
-export { RemoveTrailingSlashPlugin } from "./RemoveTrailingSlashPlugin.mjs"
